* Fixed compilation issues on 64 bit Linux * Updated version to 1.7.5
git-svn-id: http://madara.googlecode.com/svn/trunk@1324 2850746a-f1c8-11dd-b507-8b4ecd5210be

+++ include/madara/knowledge_engine/Knowledge_Record.cpp
@@ -144,7 +144,7 @@ Madara::Knowledge_Record::deep_copy (const Knowledge_Record & source)
     }
     else if (type_ == UNKNOWN_FILE_TYPE || type_ == IMAGE_JPEG)
     {
-      uint32_t temp;
+      size_t temp;
       size_ = source.size_;
       type_ = source.type_;
       file_value_ = source.to_unmanaged_buffer (temp);
